pigean-manuscripts/pyproject.toml
[project]
name = "pigean-manuscripts-overleaf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "requests>=2.31.0",
]
